Fixed iptables install and added some more tests

+++ scripts/rh-start-on-boot.sh
@@ -1,12 +1,17 @@
-# Copy rotorhazard service file
-echo "Current user is:" $USER
+# Setup Rotorhazard service
 
-sed -i -e 's/pi/'$USER'/g' ./resources/rotorhazard.service
+if [ -f "/lib/systemd/system/rotorhazard.service" ]; then
+    echo "Rotorhazard service already exsists."
+else 
+    echo "Current user is:" $USER
 
-# Note you will need to update the rotorhazard.service file if you are not using the pi username
-sudo cp ./resources/rotorhazard.service /lib/systemd/system/rotorhazard.service
-sudo chmod 644 /lib/systemd/system/rotorhazard.service
+    sed -i -e 's/pi/'$USER'/g' ./resources/rotorhazard.service
 
-# reload the daemon and enable
-sudo systemctl daemon-reload
-sudo systemctl enable rotorhazard.service
+    # Note you will need to update the rotorhazard.service file if you are not using the pi username
+    sudo cp ./resources/rotorhazard.service /lib/systemd/system/rotorhazard.service
+    sudo chmod 644 /lib/systemd/system/rotorhazard.service
+
+    # reload the daemon and enable
+    sudo systemctl daemon-reload
+    sudo systemctl enable rotorhazard.service
+fi
